+++ src/pySciCam/chronos14_raw.pyx
@@ -52,10 +52,10 @@ ctypedef np.uint16_t DTYPE_t
 @cython.wraparound(False)
 @cython.nonecheck(False)
 def read_chronos_raw(filename, int width, int height, tuple frames=None,\
-                     int bits_per_pixel=12, long start_offset = 0, int quiet = 0,\
+                     int bits_per_pixel=12, long long start_offset = 0, int quiet = 0,\
                      int old_packing_order = 0):
 
-    cdef long t0 = time.time()
+    cdef double t0 = time.time()
     cdef double bytes_per_pixel = bits_per_pixel/8.0
 
     # width and height MUST be specified.
@@ -64,7 +64,7 @@ def read_chronos_raw(filename, int width, int height, tuple frames=None,\
 
     # Get size of binary file before we start reading
     # (it might be smaller than we think)
-    cdef long nbytes = os.path.getsize(filename)
+    cdef long long nbytes = os.path.getsize(filename)
 
     # Scanlines are padded to nearest 16 bytes
     cdef unsigned int scanline_pad = 0 #int((width*1.5)%16)  # disabled for debugging
@@ -99,8 +99,8 @@ def read_chronos_raw(filename, int width, int height, tuple frames=None,\
 
 
     # Limit number of frames loaded from file (good for testing)
-    cdef long start = start_offset # bytes
-    cdef long end = nbytes
+    cdef long long start = start_offset # bytes
+    cdef long long end = nbytes
     if frames is not None:
         if frames[1] > 0:
             start = start_offset + int(bytes_per_frame*frames[0])
